src/main/java/seedu/address/model/person/Address.java
package seedu.address.model.person;

import static java.util.Objects.requireNonNull;
import static seedu.address.commons.util.AppUtil.checkArgument;

/**
 * Represents a Person's address in the address book.
 * Guarantees: immutable; is valid as declared in {@link #isValidAddress(String)}
 */
public class Address {

    public static final String MESSAGE_CONSTRAINTS =
            "Address Format: [Block Number] [Street Name]#[Unit Level]-[Unit No.]"
                    + " [Building Name] Singapore [Postal Code]";

    /*
     * E.g. 125 Orchard Road #12-34 ABC Building Singapore 123456
     */
    public static final String VALIDATION_REGEX =
            "^\\d*\\s*+[A-Za-z0-9 ]+\\s*(#\\d+-\\d+)?\\s+[A-Za-z0-9 ]+\\s+Singapore\\s+\\d{6}$";

    public final String value;

    /**
     * Constructs an {@code Address}.
     *
     * @param address A valid address.
     */
    public Address(String address) {
        requireNonNull(address);
        checkArgument(isValidAddress(address), MESSAGE_CONSTRAINTS);
        value = address;
    }

    /**
     * Returns true if a given string is a valid address.
     */
    public static boolean isValidAddress(String test) {
        return test.matches(VALIDATION_REGEX);
    }

    @Override
    public String toString() {
        return value;
    }

    @Override
    public boolean equals(Object other) {
        if (other == this) {
            return true;
        }

        // instanceof handles nulls
        if (!(other instanceof Address otherAddress)) {
            return false;
        }

        return value.equals(otherAddress.value);
    }

    @Override
    public int hashCode() {
        return value.hashCode();
    }

}
